refactor: Update profile page to dynamically generate avatar text based on first and last name
This commit modifies the profile page to dynamically generate the avatar text based on the user's first and last name. When the first name and last name inputs are changed, the avatar text is updated accordingly. If either the first name or last name is empty, the default avatar text "CN" is displayed. This enhancement improves the personalization of the profile page and provides a more customized user experience.