refactor: Add handleProfileRoutes function to middlewareHelpers.ts
This commit adds a new function called handleProfileRoutes to the middlewareHelpers.ts file. This function is responsible for handling profile routes in the middleware. It checks the validity of the JWT and redirects the user to the appropriate profile page based on the pathname. If the pathname is "/profile" and the JWT is valid, the user is redirected to "/profile/{email}". If the pathname starts with "/profile" but does not match the expected format and the JWT is valid, the user is redirected to "/profile/{email}". If the pathname matches the expected format but does not match the current user's profile, the user is redirected to "/profile/{email}". If none of these conditions are met, the middleware proceeds to the next request.